Move checkout extensions to separate file
